+++ modules/promo-codes/routes/admin.php
@@ -15,4 +15,5 @@ Route::get('/items/{server}/{item}', [AdminPromoCodeController::class, 'itemPrev
 Route::get('/{promoCode}/edit', [AdminPromoCodeController::class, 'edit'])->whereNumber('promoCode')->name('edit');
 Route::put('/{promoCode}', [AdminPromoCodeController::class, 'update'])->whereNumber('promoCode')->name('update');
 Route::patch('/{promoCode}/toggle', [AdminPromoCodeController::class, 'toggle'])->whereNumber('promoCode')->name('toggle');
+Route::patch('/{promoCode}/restore', [AdminPromoCodeController::class, 'restore'])->whereNumber('promoCode')->name('restore');
 Route::delete('/{promoCode}', [AdminPromoCodeController::class, 'destroy'])->whereNumber('promoCode')->name('destroy');